+++ test/mix/tasks/phoenix/pow.phoenix.install_test.exs
@@ -66,7 +66,9 @@ defmodule Mix.Tasks.Pow.Phoenix.InstallTest do
       Install.run(options)
 
       assert_received {:mix_shell, :info, ["Pow has been installed in your phoenix app!" <> msg]}
-      assert msg =~ "plug Pow.Plug.Session, otp_app: :test"
+      assert msg =~ "config :pow, :pow,"
+      assert msg =~ "user: Test.Users.User,"
+      assert msg =~ "plug Pow.Plug.Session, otp_app: :pow"
     end)
   end
 
@@ -86,6 +88,48 @@ defmodule Mix.Tasks.Pow.Phoenix.InstallTest do
         assert_raise Mix.Error, "mix pow.phoenix.install can only be run inside an application directory that has :phoenix as dependency", fn ->
           Install.run([])
         end
+      end)
+    end)
+  end
+
+  test "uses web app inside Phoenix umbrella app" do
+    options = @options ++ ~w(--templates --extension PowResetPassword --extension PowEmailConfirmation)
+    File.cd!(@tmp_path, fn ->
+      File.write!("mix.exs", """
+      defmodule MyAppWeb.MixProject do
+        use Mix.Project
+
+        def project do
+          [
+            app: :my_app_web,
+            deps: [
+              {:phoenix, ">= 0.0.0"}
+            ]
+          ]
+        end
+      end
+      """)
+
+      Application.put_env(:my_app_web, :generators, context_app: :my_app)
+
+      Mix.Project.in_project(:my_app_web, ".", fn _ ->
+        Install.run(options)
+
+        assert_received {:mix_shell, :info, ["Pow has been installed in your phoenix app!" <> msg]}
+        assert msg =~ "config :my_app_web, :pow,"
+        assert msg =~ "user: MyApp.Users.User,"
+        assert msg =~ "plug Pow.Plug.Session, otp_app: :my_app_web"
+
+        assert_received {:mix_shell, :info, ["Pow Phoenix templates and views has been generated." <> msg]}
+        assert msg =~ "repo: MyApp.Repo"
+        assert msg =~ "user: MyApp.Users.User"
+        assert msg =~ "web_module: MyAppWeb"
+
+        assert File.exists?(Path.join(["lib", "my_app_web", "templates", "pow"]))
+        assert File.exists?(Path.join(["lib", "my_app_web", "views", "pow"]))
+
+        assert File.exists?(Path.join(["lib", "my_app_web", "templates", "pow_reset_password"]))
+        assert File.exists?(Path.join(["lib", "my_app_web", "views", "pow_reset_password"]))
       end)
     end)
   end