+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.0.4 (TBA)
 
 * Added `PowInvitation` to the `mix pow.extension.phoenix.gen.templates` and `mix pow.extension.phoenix.mailer.gen.templates` tasks
+* Fixed issue in umbrella projects where extensions wasn't found in environment configuration
 
 ## v1.0.3 (2019-03-09)